# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  models_weather.CurrentWeather:
    properties:
      temperature:
        type: number
    type: object
  models_weather.DailyForecast:
    properties:
      date:
        type: string
      description:
        type: string
      max_temperature:
        type: number
      min_temperature:
        type: number
      temperature:
        type: number
    type: object
  models_weather.Location:
    properties:
      lat:
        type: number
      lon:
        type: number
    type: object
  models_weather.LocationWeather:
    properties:
      error:
        type: string
      lat:
        type: number
      lon:
        type: number
      temperature:
        type: number
    type: object
  shared_responses.StatusResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  shared_responses.SuccessResponse-array_models_weather_DailyForecast:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models_weather.DailyForecast'
        type: array
      message:
        type: string
    type: object
  shared_responses.SuccessResponse-array_models_weather_LocationWeather:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models_weather.LocationWeather'
        type: array
      message:
        type: string
    type: object
  shared_responses.SuccessResponse-models_weather_CurrentWeather:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models_weather.CurrentWeather'
      message:
        type: string
    type: object
info:
  contact: {}
  title: Wheather Example API
  version: "1.0"
paths:
  /weather:
    get:
      description: Returns current weather for given coordinates
      parameters:
      - default: "18.300231990440125"
        description: Latitude
        in: query
        name: lat
        required: true
        type: string
      - default: "-64.8251590359234"
        description: Longitude
        in: query
        name: lon
        required: true
        type: string
      - description: Weather provider
        enum:
        - openweather
        - openmeteo
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared_responses.SuccessResponse-models_weather_CurrentWeather'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_responses.StatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared_responses.StatusResponse'
      summary: Get Current Weather
      tags:
      - weather
  /weather/forecast:
    get:
      description: Returns weather forecast for 5 days
      parameters:
      - default: "53.9"
        description: Latitude
        in: query
        name: lat
        required: true
        type: string
      - default: "27.5667"
        description: Longitude
        in: query
        name: lon
        required: true
        type: string
      - description: Weather provider
        enum:
        - openweather
        - openmeteo
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared_responses.SuccessResponse-array_models_weather_DailyForecast'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_responses.StatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared_responses.StatusResponse'
      summary: Get Weather Forecast
      tags:
      - weather
  /weather/multiple:
    post:
      consumes:
      - application/json
      description: Returns current temperature for multiple coordinates
      parameters:
      - description: Array of locations
        in: body
        name: locations
        required: true
        schema:
          items:
            $ref: '#/definitions/models_weather.Location'
          type: array
      - description: Weather provider
        enum:
        - openweather
        - openmeteo
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared_responses.SuccessResponse-array_models_weather_LocationWeather'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_responses.StatusResponse'
      summary: Get Weather for Multiple Locations
      tags:
      - weather
swagger: "2.0"